refactor: updating import paths and devDeps
