+++ packages/insights-common-typescript/package.json
@@ -25,7 +25,7 @@
   "peerDependencies": {
     "@patternfly/react-core": "^4.128.2",
     "@patternfly/react-icons": "^4.10.11",
-    "@redhat-cloud-services/frontend-components": "^3.2.8",
+    "@redhat-cloud-services/frontend-components": "^3.8.12",
     "@redhat-cloud-services/frontend-components-notifications": "^3.2.2",
     "@redhat-cloud-services/frontend-components-utilities": "^3.2.2",
     "@redhat-cloud-services/rbac-client": "^1.0.98",
