+++ packages/insights-common-typescript/src/components/ErrorBoundaryPage/ErrorBoundaryPage.tsx
@@ -4,7 +4,7 @@ import { Messages } from '../../properties/Messages';
 import { ErrorCircleOIcon } from '@patternfly/react-icons';
 import { Text, ExpandableSection } from '@patternfly/react-core';
 import { global_spacer_sm, global_BackgroundColor_dark_300 } from '@patternfly/react-tokens';
-import { join } from '../../utils/ComponentUtils';
+import { join } from '../../utils';
 import { style } from 'typestyle';
 import { EmptyState } from '../EmptyState/EmptyState';
 import { ouiaParams } from '../../utils/Ouia';
