+++ packages/insights-common-typescript/src/components/ErrorBoundaryPage/__tests__/ErrorBoundaryPage.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { ErrorBoundaryPage } from '../../..';
+import { ErrorBoundaryPage } from '../ErrorBoundaryPage';
 import jestMock from 'jest-mock';
 
 jest.mock('@redhat-cloud-services/frontend-components', () => {
